LPS-148315 Using a parent component for AnalyticsReports with handlers for the user events
